Add codes-example.txt
This is an example of what `codes.txt` should look like. This is already outlined in the README, but I feel like this is akin to env vars which I like to provide a base template for, even though this isn't as strict as key=values that env files are structured like.